feat: delete Timer

+++ src/storage/storageTimerControl.ts
@@ -6,10 +6,25 @@ import { AppTimer } from "@contexts/TimerControlContext";
 
 export async function timerControlStorageSave(app: AppTimer) {
   const storedApps = await timerControlGetApps();
+  const isAlreadyExist = storedApps.filter(timer => timer.appName !== app.appName)
+
+  if (isAlreadyExist) {
+    const storage = JSON.stringify([...isAlreadyExist, app])
+    await AsyncStorage.setItem(APPS_TIMER_CONTROL, storage)
+  } else {
+    const storage = JSON.stringify([...storedApps, app])
+    await AsyncStorage.setItem(APPS_TIMER_CONTROL, storage)
+  }
 
 
-  const storage = JSON.stringify([...storedApps, app]);
-  await AsyncStorage.setItem(APPS_TIMER_CONTROL, storage)
+}
+
+
+
+export async function timerControlStorageRemoveSpecificAppTimer(app: AppTimer) {
+  const storedApps = await timerControlGetApps();
+  const filtered = storedApps.filter(timer => timer.appName !== app.appName)
+  await AsyncStorage.setItem(APPS_TIMER_CONTROL, JSON.stringify(filtered))
 }
